backend/requirements.txt
fastapi==0.128.4
uvicorn[standard]==0.39.0
openai==2.15.0
requests==2.32.5
python-dotenv>=1.0.0
sqlmodel==0.0.32
pydantic==2.12.5
urllib3<2.0.0
langgraph>=0.0.30
langchain-core>=0.1.40
langchain-openai>=0.1.3
